Move OBJ saving logic into the converter
